% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plots}
\alias{plots}
\title{Plots}
\usage{
plots(Response, Predictors, pl_type = "hist")
}
\arguments{
\item{Response}{A \code{data-frame} contains the response value.}

\item{Predictors}{A \code{data-frame} contains the predictor values.}

\item{pl_type}{A \code{char} desired plot type.}
}
\value{
A \code{data frame} c("res_fit", "qq", "hist, all"), res_fit -> Residuals vs fitted-values, qq -> qq-plot of residuals, hist -> Histogram of residuals, all -> all of the plots together
\describe{
     }
}
\description{
Residuals vs fitted-values, qq-plot of residuals, Histogram of residuals | Plots
}
\examples{
ci(Response, Predictors, pl_type = "hist")
}
\author{
Ozan Turkes, Ayomide Afolabi, Geeta Kharel
}
