% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_function.R
\name{pval}
\alias{pval}
\title{F-Test}
\usage{
pval(Response, Predictors)
}
\arguments{
\item{Response}{A \code{data-frame} containing the response value in the dataset.}

\item{Predictors}{A \code{data-frame} containinf the different type of predictors in the dataset.}
}
\value{
A \code{data frame} containing the following attributes:
\describe{
     \item{p_value}{Estimated p value}
     }
}
\description{
Estimate the coefficient vector
}
\examples{
coeff(Response, Predictors)
}
\author{
Ayomide Afolabi, Ozan Turkes, Geeta Kharel
}
